Changed back to Integer casts of criteria.uniqueResult() (Hibernate version 3.3.2.GA returns Integer)
git-svn-id: https://ark-informatics.googlecode.com/svn/ARK@6467 2beaffb7-5388-1629-0768-26d37724f1fc

+++ ark-lims/src/main/java/au/org/theark/lims/model/dao/BioCollectionDao.java
@@ -231,7 +231,7 @@ public class BioCollectionDao extends HibernateSessionDao implements IBioCollect
 		}
 		Criteria criteria = buildBioCollectionCriteria(bioCollectionCriteria);
 		criteria.setProjection(Projections.rowCount());
-		Long totalCount = (Long) criteria.uniqueResult();
+		Integer totalCount = (Integer) criteria.uniqueResult();
 		return totalCount.intValue();
 	}
